履歴書構成データのファイルの分割。
